requirements.txt
pycodestyle
pytest
pygame
pylint
pyyaml